// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Int or string */
  ActionCount: any;
  /** AreaOfEffectType ("SPHERE", "CUBE", "CYLINDER", "LINE", "CONE") or list of AreaOfEffectTypes */
  AreaOfEffectTypeFilter: any;
  /** Float, list of floats, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  FloatFilter: any;
  /** Int, list of ints, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  IntFilter: any;
  /** LanguageScript ("COMMON", "ELVISH", "DWARVISH", "INFERNAL", "DRACONIC", "CELESTIAL") or list of LanguageScripts */
  LanguageScriptFilter: any;
  /** MonsterSubtype ("ANY_RACE", "HUMAN", "DWARF", "ELF", "GOBLINOID", "MERFOLK", "SHAPECHANGER", "DEMON", "DEVIL", "ORC", "SAHUAGIN", "TITAN", "KOBOLD", "GNOLL", "GRIMLOCK", "LIZARDFOLK", "GNOME") or list of MonsterSubtypes */
  MonsterSubtypeFilter: any;
  /** MonsterTypes ("BEAST", "MONSTROSITY", "DRAGON", "HUMANOID", "UNDEAD", "FIEND", "CELESTIAL", "CONSTRUCT", "GIANT", "ELEMENTAL", "FEY", "ABERRATION", "OOZE", "SWARM", "PLANT") or list of MonsterTypes */
  MonsterTypeFilter: any;
  /** ProficiencyType ("WEAPONS", "ARTISANS_TOOLS", "SKILLS", "ARMOR", "MUSICAL_INSTRUMENTS", "SAVING_THROWS", "OTHER", "GAMING_SETS", "VEHICLES") or list of ProficiencyTypes */
  ProficiencyTypeFilter: any;
  /** Size ("TINY", "SMALL", "MEDIUM", "LARGE", "HUGE", "GARGANTUAN") or list of sizes */
  SizeFilter: any;
  /** SpellAttackType ("MELEE", "RANGED") or list of SpellAttackTypes */
  SpellAttackTypeFilter: any;
  /** String or list of strings */
  StringFilter: any;
};
// ...
export type Cost = {
  __typename?: 'Cost';
  quantity: Scalars['Int'];
  unit: Currency;
};

export type CountedReferenceOption = {
  __typename?: 'CountedReferenceOption';
  count: Scalars['Int'];
  of: IEquipment;
  option_type: Scalars['String'];
  prerequisites?: Maybe<Array<ProficiencyPrerequisite>>;
};

export enum Currency {
  Cp = 'CP',
  Gp = 'GP',
  Sp = 'SP'
}

export type Damage = {
  __typename?: 'Damage';
  damage_dice: Scalars['String'];
  damage_type: DamageType;
};
// ...
export type DamageType = {
  __typename?: 'DamageType';
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type EquipmentCategory = {
  __typename?: 'EquipmentCategory';
  equipment: Array<IEquipmentBase>;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type IEquipment = {
  cost: Cost;
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
  weight?: Maybe<Scalars['Float']>;
};

export type IEquipmentBase = {
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Range = {
  __typename?: 'Range';
  long?: Maybe<Scalars['Int']>;
  normal: Scalars['Int'];
};
// ...
export type Weapon = IEquipment & IEquipmentBase & {
  __typename?: 'Weapon';
  category_range: EquipmentCategory;
  cost: Cost;
  damage?: Maybe<Damage>;
  desc?: Maybe<Array<Scalars['String']>>;
  equipment_category: EquipmentCategory;
  index: Scalars['String'];
  name: Scalars['String'];
  properties: Array<WeaponProperty>;
  range: Range;
  special?: Maybe<Array<Scalars['String']>>;
  throw_range?: Maybe<Range>;
  two_handed_damage?: Maybe<Damage>;
  weapon_category: EquipmentCategory;
  weapon_range: WeaponRange;
  weight?: Maybe<Scalars['Float']>;
};


export type WeaponPropertiesArgs = {
  name?: InputMaybe<Scalars['String']>;
};

export type WeaponProperty = {
  __typename?: 'WeaponProperty';
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};

export enum WeaponRange {
  Melee = 'MELEE',
  Ranged = 'RANGED'
}